+++ packages/ubuntu_desktop_installer/lib/pages/active_directory/active_directory_l10n.dart
@@ -1,5 +1,5 @@
 import 'package:ubuntu_desktop_installer/l10n.dart';
-import 'package:ubuntu_desktop_installer/services.dart';
+import 'package:ubuntu_provision/services.dart';
 
 extension AdDomainNameValidationL10n on AdDomainNameValidation {
   String localize(AppLocalizations lang) {
