+++ packages/ubuntu_desktop_installer/lib/pages/active_directory/active_directory_model.dart
@@ -3,6 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:safe_change_notifier/safe_change_notifier.dart';
 import 'package:ubuntu_desktop_installer/services.dart';
+import 'package:ubuntu_provision/services.dart';
 
 final activeDirectoryModelProvider = ChangeNotifierProvider(
   (_) => ActiveDirectoryModel(getService<ActiveDirectoryService>()),
