+++ packages/ubuntu_desktop_installer/test/active_directory/test_active_directory.mocks.dart
@@ -9,7 +9,7 @@ import 'dart:ui' as _i5;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:ubuntu_desktop_installer/pages/active_directory/active_directory_model.dart'
     as _i2;
-import 'package:ubuntu_desktop_installer/services.dart' as _i4;
+import 'package:ubuntu_provision/services.dart' as _i4;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
