+++ packages/ubuntu_desktop_installer/test/services/active_directory_service_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:subiquity_test/subiquity_test.dart';
 import 'package:ubuntu_desktop_installer/services/active_directory_service.dart';
+import 'package:ubuntu_provision/src/services/active_directory_service.dart';
 
 void main() {
   group('subiquity', () {
