+++ packages/ubuntu_welcome/lib/services.dart
@@ -1,3 +1,4 @@
+export 'services/realmd_active_directory_service.dart';
 export 'services/xdg_identity_service.dart';
 export 'services/xdg_keyboard_service.dart';
 export 'services/xdg_locale_service.dart';
